Silenced a SafeMalloc Formatting warning with an explicit cast to (long).
-RL


git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@1705 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/core/misc.c
@@ -183,7 +183,7 @@ GENERIC_PTR SafeMalloc(size_t amount)
     
     /* Mando's Fix is to put a ckalloc here */
     if((ptr = malloc(amount)) == NULL) {
-        printf("Error: SafeMalloc could not allocate the requested %lu bytes\n",amount);
+        printf("Error: SafeMalloc could not allocate the requested %lu bytes\n",(long) amount);
         ExitStageRight();
         exit(0);
     }
